Move mongo.js to utils folder
